feat: refactor code structure and improve module imports for clarity
