+++ tools/update_swagger_ui.py
@@ -22,7 +22,7 @@ from pathlib import Path
 
 import requests
 
-from apispec_aiohttp.aiohttp import SWAGGER_UI_STATIC_FILES, SWAGGER_UI_VERSION_PATH
+from apispec_aiohttp.swagger_ui import SWAGGER_UI_STATIC_FILES, SWAGGER_UI_VERSION_PATH
 
 # Configure logging
 logging.basicConfig(
